requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
psycopg2-binary>=2.9
dj-database-url>=2.1
gunicorn>=21.2

boto3>=1.34
python-json-logger>=2.0
django-filter>=23.5
drf-yasg>=1.21
watchtower>=3.0
djangorestframework-simplejwt>=5.3
bleach>=6.0
celery>=5.3
django-redis>=5.2
django-prometheus>=2.3

